requirements.txt
PySide6==6.6.1
psutil==5.9.8
requests==2.31.0
pyserial==3.5
wmi==1.5.1
